EDU-583 Rename urls module to routes
