+++ src/components/document-info-cell.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import urls from '../utils/urls.js';
+import urls from '../utils/routes.js';
 import { useTranslation } from 'react-i18next';
 import { useDateFormat } from './locale-context.js';
 import { documentMetadataShape } from '../ui/default-prop-types.js';
