+++ src/components/pages/batches.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import { Table, List } from 'antd';
-import urls from '../../utils/urls.js';
+import urls from '../../utils/routes.js';
 import Logger from '../../common/logger.js';
 import { useTranslation } from 'react-i18next';
 import ImportTypeIcon from '../import-type-icon.js';
