+++ src/components/pages/doc.js
@@ -1,6 +1,6 @@
 import { message } from 'antd';
 import PropTypes from 'prop-types';
-import urls from '../../utils/urls.js';
+import urls from '../../utils/routes.js';
 import Restricted from '../restricted.js';
 import Logger from '../../common/logger.js';
 import { useUser } from '../user-context.js';
