+++ src/components/pages/docs.js
@@ -1,7 +1,7 @@
 import by from 'thenby';
 import Table from '../table.js';
 import PropTypes from 'prop-types';
-import urls from '../../utils/urls.js';
+import urls from '../../utils/routes.js';
 import Restricted from '../restricted.js';
 import Logger from '../../common/logger.js';
 import { useUser } from '../user-context.js';
