+++ src/components/pages/imports.js
@@ -1,7 +1,7 @@
 import by from 'thenby';
 import React from 'react';
 import PropTypes from 'prop-types';
-import urls from '../../utils/urls.js';
+import urls from '../../utils/routes.js';
 import { useTranslation } from 'react-i18next';
 import { Table, Collapse, Button } from 'antd';
 import { useDateFormat } from '../locale-context.js';
