+++ src/components/pages/room.js
@@ -1,7 +1,7 @@
 import by from 'thenby';
 import PropTypes from 'prop-types';
 import Markdown from '../markdown.js';
-import urls from '../../utils/urls.js';
+import urls from '../../utils/routes.js';
 import Logger from '../../common/logger.js';
 import { useUser } from '../user-context.js';
 import FavoriteStar from '../favorite-star.js';
