+++ src/server/document-controller.js
@@ -1,6 +1,7 @@
 import express from 'express';
-import urls from '../utils/urls.js';
 import httpErrors from 'http-errors';
+import routes from '../utils/routes.js';
+import urlUtils from '../utils/url-utils.js';
 import PageRenderer from './page-renderer.js';
 import { PAGE_NAME } from '../domain/page-name.js';
 import DocumentService from '../services/document-service.js';
@@ -48,7 +49,7 @@ class DocumentController {
     const { user } = req;
     const { docKey } = req.params;
     const { view, templateDocumentKey } = req.query;
-    const routeWildcardValue = urls.removeLeadingSlash(req.params[0]);
+    const routeWildcardValue = urlUtils.removeLeadingSlash(req.params[0]);
 
     const doc = await this.documentService.getDocumentByKey(docKey);
     if (!doc) {
@@ -56,13 +57,13 @@ class DocumentController {
     }
 
     if (doc.slug !== routeWildcardValue) {
-      return res.redirect(301, urls.getDocUrl({ key: doc.key, slug: doc.slug, view, templateDocumentKey }));
+      return res.redirect(301, routes.getDocUrl({ key: doc.key, slug: doc.slug, view, templateDocumentKey }));
     }
 
     let templateDocument;
     if (templateDocumentKey) {
       if (doc.sections.length) {
-        return res.redirect(302, urls.getDocUrl({ key: doc.key, slug: doc.slug }));
+        return res.redirect(302, routes.getDocUrl({ key: doc.key, slug: doc.slug }));
       }
 
       templateDocument = await this.documentService.getDocumentByKey(templateDocumentKey);
