+++ src/server/storage-controller.js
@@ -1,7 +1,7 @@
 import os from 'os';
 import multer from 'multer';
 import express from 'express';
-import urls from '../utils/urls.js';
+import urls from '../utils/routes.js';
 import httpErrors from 'http-errors';
 import prettyBytes from 'pretty-bytes';
 import permissions from '../domain/permissions.js';
