+++ src/server/user-controller.js
@@ -1,7 +1,7 @@
 /* eslint-disable max-params */
 import express from 'express';
 import passport from 'passport';
-import urls from '../utils/urls.js';
+import urls from '../utils/routes.js';
 import httpErrors from 'http-errors';
 import session from 'express-session';
 import MongoStore from 'connect-mongo';
