+++ src/utils/url-utils.js
@@ -1,0 +1,12 @@
+function removeTrailingSlash(path) {
+  return String(path).replace(/\/*$/, '');
+}
+
+function removeLeadingSlash(path) {
+  return String(path).replace(/^\/*/, '');
+}
+
+export default {
+  removeTrailingSlash,
+  removeLeadingSlash
+};
