+++ src/utils/url-utils.spec.js
@@ -1,0 +1,66 @@
+import {
+  removeLeadingSlash,
+  removeTrailingSlash
+} from './url-utils.js';
+
+describe('url-utils', () => {
+  let result;
+
+  describe('removeTrailingSlash', () => {
+    const testCases = [
+      {
+        path: '/some-path/some-other-path',
+        expectedResult: '/some-path/some-other-path'
+      },
+      {
+        path: '/some-path/some-other-path/',
+        expectedResult: '/some-path/some-other-path'
+      },
+      {
+        path: '/some-path/some-other-path///',
+        expectedResult: '/some-path/some-other-path'
+      }
+    ];
+
+    testCases.forEach(({ path, expectedResult }) => {
+      describe(`when path is '${path}'`, () => {
+        beforeEach(() => {
+          result = removeTrailingSlash(path);
+        });
+        it(`should return '${expectedResult}'`, () => {
+          expect(result).toBe(expectedResult);
+        });
+      });
+    });
+
+  });
+
+  describe('removeLeadingSlash', () => {
+
+    const testCases = [
+      {
+        path: 'some-path/some-other-path/',
+        expectedResult: 'some-path/some-other-path/'
+      },
+      {
+        path: '/some-path/some-other-path/',
+        expectedResult: 'some-path/some-other-path/'
+      },
+      {
+        path: '///some-path/some-other-path/',
+        expectedResult: 'some-path/some-other-path/'
+      }
+    ];
+
+    testCases.forEach(({ path, expectedResult }) => {
+      describe(`when path is '${path}'`, () => {
+        beforeEach(() => {
+          result = removeLeadingSlash(path);
+        });
+        it(`it should return '${expectedResult}'`, () => {
+          expect(result).toBe(expectedResult);
+        });
+      });
+    });
+  });
+});
